requirements.txt
# Core ML/CV Dependencies
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.3.0
opencv-python>=4.6.0
numpy>=1.23.0

# SLAM Dependencies  
scipy>=1.4.1
matplotlib>=3.3.0
pillow>=7.1.2

# Sensor Fusion & Utils
pyyaml>=5.3.1
tqdm>=4.64.0
pandas>=1.1.4
seaborn>=0.11.0

# Development & Testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
pre-commit>=2.15.0

# Deployment
docker>=6.0.0
